INTYGFV-14680: Refactor convert code for internal <-> certificate. Break convert toolkits to separate implementation classes for each question, category and metadata.
