+++ af/af00213/src/main/java/se/inera/intyg/common/af00213/v1/model/converter/certificate/CategoryAktivitetsbegransning.java
@@ -1,0 +1,36 @@
+package se.inera.intyg.common.af00213.v1.model.converter.certificate;
+
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.AKTIVITETSBEGRANSNING_CATEGORY_ID;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.AKTIVITETSBEGRANSNING_CATEGORY_TEXT;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.FUNKTIONSNEDSATTNING_DELSVAR_ID_11;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.FUNKTIONSNEDSATTNING_SVAR_JSON_ID_11;
+import static se.inera.intyg.common.support.facade.util.ValidationExpressionToolkit.singleExpression;
+
+import se.inera.intyg.common.services.texts.CertificateTextProvider;
+import se.inera.intyg.common.support.facade.model.CertificateDataElement;
+import se.inera.intyg.common.support.facade.model.config.CertificateDataConfigCategory;
+import se.inera.intyg.common.support.facade.model.validation.CertificateDataValidation;
+import se.inera.intyg.common.support.facade.model.validation.CertificateDataValidationShow;
+
+public class CategoryAktivitetsbegransning {
+
+    public static CertificateDataElement toCertificate(int index, CertificateTextProvider texts) {
+        return CertificateDataElement.builder()
+            .id(AKTIVITETSBEGRANSNING_CATEGORY_ID)
+            .index(index)
+            .config(
+                CertificateDataConfigCategory.builder()
+                    .text(texts.get(AKTIVITETSBEGRANSNING_CATEGORY_TEXT))
+                    .build()
+            )
+            .validation(
+                new CertificateDataValidation[]{
+                    CertificateDataValidationShow.builder()
+                        .questionId(FUNKTIONSNEDSATTNING_DELSVAR_ID_11)
+                        .expression(singleExpression(FUNKTIONSNEDSATTNING_SVAR_JSON_ID_11))
+                        .build()
+                }
+            )
+            .build();
+    }
+}
