+++ af/af00213/src/main/java/se/inera/intyg/common/af00213/v1/model/converter/certificate/CategoryUtredningBehandling.java
@@ -1,0 +1,23 @@
+package se.inera.intyg.common.af00213.v1.model.converter.certificate;
+
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.UTREDNING_BEHANDLING_CATEGORY_ID;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.UTREDNING_BEHANDLING_CATEGORY_TEXT;
+
+import se.inera.intyg.common.services.texts.CertificateTextProvider;
+import se.inera.intyg.common.support.facade.model.CertificateDataElement;
+import se.inera.intyg.common.support.facade.model.config.CertificateDataConfigCategory;
+
+public class CategoryUtredningBehandling {
+
+    public static CertificateDataElement toCertificate(int index, CertificateTextProvider texts) {
+        return CertificateDataElement.builder()
+            .id(UTREDNING_BEHANDLING_CATEGORY_ID)
+            .index(index)
+            .config(
+                CertificateDataConfigCategory.builder()
+                    .text(texts.get(UTREDNING_BEHANDLING_CATEGORY_TEXT))
+                    .build()
+            )
+            .build();
+    }
+}
