+++ af/af00213/src/main/java/se/inera/intyg/common/af00213/v1/model/converter/certificate/QuestionAktivitetsbegransning.java
@@ -1,0 +1,65 @@
+package se.inera.intyg.common.af00213.v1.model.converter.certificate;
+
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.AKTIVITETSBEGRANSNING_DELSVAR_DESCRIPTION;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.AKTIVITETSBEGRANSNING_DELSVAR_ID_21;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.AKTIVITETSBEGRANSNING_DELSVAR_ID_22;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.AKTIVITETSBEGRANSNING_DELSVAR_TEXT;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.AKTIVITETSBEGRANSNING_SVAR_JSON_ID_21;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.AKTIVITETSBEGRANSNING_SVAR_JSON_ID_22;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.FUNKTIONSNEDSATTNING_DELSVAR_ID_11;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.FUNKTIONSNEDSATTNING_SVAR_JSON_ID_11;
+import static se.inera.intyg.common.support.facade.util.ValidationExpressionToolkit.singleExpression;
+import static se.inera.intyg.common.support.facade.util.ValueToolkit.textValue;
+
+import se.inera.intyg.common.services.texts.CertificateTextProvider;
+import se.inera.intyg.common.support.facade.model.Certificate;
+import se.inera.intyg.common.support.facade.model.CertificateDataElement;
+import se.inera.intyg.common.support.facade.model.config.CertificateDataConfigTextArea;
+import se.inera.intyg.common.support.facade.model.validation.CertificateDataValidation;
+import se.inera.intyg.common.support.facade.model.validation.CertificateDataValidationMandatory;
+import se.inera.intyg.common.support.facade.model.validation.CertificateDataValidationShow;
+import se.inera.intyg.common.support.facade.model.value.CertificateDataTextValue;
+
+public class QuestionAktivitetsbegransning {
+
+    public static CertificateDataElement toCertificate(String aktivitetsbegransning, int index, CertificateTextProvider texts) {
+        return CertificateDataElement.builder()
+            .id(AKTIVITETSBEGRANSNING_DELSVAR_ID_22)
+            .parent(AKTIVITETSBEGRANSNING_DELSVAR_ID_21)
+            .index(index)
+            .config(
+                CertificateDataConfigTextArea.builder()
+                    .text(texts.get(AKTIVITETSBEGRANSNING_DELSVAR_TEXT))
+                    .description(texts.get(AKTIVITETSBEGRANSNING_DELSVAR_DESCRIPTION))
+                    .id(AKTIVITETSBEGRANSNING_SVAR_JSON_ID_22)
+                    .build()
+            )
+            .value(
+                CertificateDataTextValue.builder()
+                    .id(AKTIVITETSBEGRANSNING_SVAR_JSON_ID_22)
+                    .text(aktivitetsbegransning)
+                    .build()
+            )
+            .validation(
+                new CertificateDataValidation[]{
+                    CertificateDataValidationMandatory.builder()
+                        .questionId(AKTIVITETSBEGRANSNING_DELSVAR_ID_22)
+                        .expression(singleExpression(AKTIVITETSBEGRANSNING_SVAR_JSON_ID_22))
+                        .build(),
+                    CertificateDataValidationShow.builder()
+                        .questionId(FUNKTIONSNEDSATTNING_DELSVAR_ID_11)
+                        .expression(singleExpression(FUNKTIONSNEDSATTNING_SVAR_JSON_ID_11))
+                        .build(),
+                    CertificateDataValidationShow.builder()
+                        .questionId(AKTIVITETSBEGRANSNING_DELSVAR_ID_21)
+                        .expression(singleExpression(AKTIVITETSBEGRANSNING_SVAR_JSON_ID_21))
+                        .build()
+                }
+            )
+            .build();
+    }
+
+    public static String toInternal(Certificate certificate) {
+        return textValue(certificate.getData(), AKTIVITETSBEGRANSNING_DELSVAR_ID_22, AKTIVITETSBEGRANSNING_SVAR_JSON_ID_22);
+    }
+}
