+++ af/af00213/src/main/java/se/inera/intyg/common/af00213/v1/model/converter/certificate/QuestionHarFunktionsnedsattning.java
@@ -1,0 +1,57 @@
+package se.inera.intyg.common.af00213.v1.model.converter.certificate;
+
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.ANSWER_NO;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.ANSWER_YES;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.FUNKTIONSNEDSATTNING_CATEGORY_ID;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.FUNKTIONSNEDSATTNING_DELSVAR_ID_11;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.FUNKTIONSNEDSATTNING_QUESTION_DESCRIPTION;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.FUNKTIONSNEDSATTNING_QUESTION_TEXT;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.FUNKTIONSNEDSATTNING_SVAR_JSON_ID_11;
+import static se.inera.intyg.common.support.facade.util.ValidationExpressionToolkit.singleExpression;
+import static se.inera.intyg.common.support.facade.util.ValueToolkit.booleanValue;
+
+import se.inera.intyg.common.services.texts.CertificateTextProvider;
+import se.inera.intyg.common.support.facade.model.Certificate;
+import se.inera.intyg.common.support.facade.model.CertificateDataElement;
+import se.inera.intyg.common.support.facade.model.config.CertificateDataConfigBoolean;
+import se.inera.intyg.common.support.facade.model.validation.CertificateDataValidation;
+import se.inera.intyg.common.support.facade.model.validation.CertificateDataValidationMandatory;
+import se.inera.intyg.common.support.facade.model.value.CertificateDataValueBoolean;
+
+public class QuestionHarFunktionsnedsattning {
+
+    public static CertificateDataElement toCertificate(Boolean harFunktionsnedsattning, int index, CertificateTextProvider texts) {
+        return CertificateDataElement.builder()
+            .id(FUNKTIONSNEDSATTNING_DELSVAR_ID_11)
+            .index(index)
+            .parent(FUNKTIONSNEDSATTNING_CATEGORY_ID)
+            .config(
+                CertificateDataConfigBoolean.builder()
+                    .text(texts.get(FUNKTIONSNEDSATTNING_QUESTION_TEXT))
+                    .description(texts.get(FUNKTIONSNEDSATTNING_QUESTION_DESCRIPTION))
+                    .id(FUNKTIONSNEDSATTNING_SVAR_JSON_ID_11)
+                    .selectedText(texts.get(ANSWER_YES))
+                    .unselectedText(texts.get(ANSWER_NO))
+                    .build()
+            )
+            .value(
+                CertificateDataValueBoolean.builder()
+                    .id(FUNKTIONSNEDSATTNING_SVAR_JSON_ID_11)
+                    .selected(harFunktionsnedsattning)
+                    .build()
+            )
+            .validation(
+                new CertificateDataValidation[]{
+                    CertificateDataValidationMandatory.builder()
+                        .questionId(FUNKTIONSNEDSATTNING_DELSVAR_ID_11)
+                        .expression(singleExpression(FUNKTIONSNEDSATTNING_SVAR_JSON_ID_11))
+                        .build()
+                }
+            )
+            .build();
+    }
+
+    public static Boolean toInternal(Certificate certificate) {
+        return booleanValue(certificate.getData(), FUNKTIONSNEDSATTNING_DELSVAR_ID_11, FUNKTIONSNEDSATTNING_SVAR_JSON_ID_11);
+    }
+}
