+++ af/af00213/src/main/java/se/inera/intyg/common/af00213/v1/model/converter/certificate/QuestionOvrigt.java
@@ -1,0 +1,43 @@
+package se.inera.intyg.common.af00213.v1.model.converter.certificate;
+
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.OVRIGT_CATEGORY_ID;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.OVRIGT_DELSVAR_ID_5;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.OVRIGT_QUESTION_DESCRIPTION;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.OVRIGT_QUESTION_TEXT;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.OVRIGT_SVAR_JSON_ID_5;
+import static se.inera.intyg.common.support.facade.util.ValueToolkit.textValue;
+
+import se.inera.intyg.common.services.texts.CertificateTextProvider;
+import se.inera.intyg.common.support.facade.model.Certificate;
+import se.inera.intyg.common.support.facade.model.CertificateDataElement;
+import se.inera.intyg.common.support.facade.model.config.CertificateDataConfigTextArea;
+import se.inera.intyg.common.support.facade.model.value.CertificateDataTextValue;
+
+public class QuestionOvrigt {
+
+    public static CertificateDataElement toCerticate(String ovrigt, int index,
+        CertificateTextProvider texts) {
+        return CertificateDataElement.builder()
+            .id(OVRIGT_DELSVAR_ID_5)
+            .index(index)
+            .parent(OVRIGT_CATEGORY_ID)
+            .config(
+                CertificateDataConfigTextArea.builder()
+                    .id(OVRIGT_SVAR_JSON_ID_5)
+                    .text(texts.get(OVRIGT_QUESTION_TEXT))
+                    .description(texts.get(OVRIGT_QUESTION_DESCRIPTION))
+                    .build()
+            )
+            .value(
+                CertificateDataTextValue.builder()
+                    .id(OVRIGT_SVAR_JSON_ID_5)
+                    .text(ovrigt)
+                    .build()
+            )
+            .build();
+    }
+
+    public static String toInternal(Certificate certificate) {
+        return textValue(certificate.getData(), OVRIGT_DELSVAR_ID_5, OVRIGT_SVAR_JSON_ID_5);
+    }
+}
