+++ af/af00213/src/test/java/se/inera/intyg/common/af00213/v1/model/converter/certificate/QuestionHarFunktionsnedsattningTest.java
@@ -1,0 +1,66 @@
+package se.inera.intyg.common.af00213.v1.model.converter.certificate;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.when;
+
+import java.util.stream.Stream;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import se.inera.intyg.common.af00213.v1.model.converter.CertificateToInternal;
+import se.inera.intyg.common.af00213.v1.model.internal.Af00213UtlatandeV1;
+import se.inera.intyg.common.services.texts.CertificateTextProvider;
+import se.inera.intyg.common.support.facade.builder.CertificateBuilder;
+import se.inera.intyg.common.support.model.common.internal.GrundData;
+
+@ExtendWith(MockitoExtension.class)
+class QuestionHarFunktionsnedsattningTest {
+
+    @Mock
+    private CertificateTextProvider texts;
+
+    @BeforeEach
+    void setup() {
+        when(texts.get(Mockito.any(String.class))).thenReturn("Test string");
+    }
+
+    @Nested
+    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
+    class ToCertificate {
+
+        private Af00213UtlatandeV1 internalCertificate;
+
+        @BeforeEach
+        void setup() {
+            internalCertificate = Af00213UtlatandeV1.builder()
+                .setGrundData(new GrundData())
+                .setId("id")
+                .setTextVersion("TextVersion")
+                .build();
+        }
+
+        Stream<Boolean> booleanValues() {
+            return Stream.of(true, false, null);
+        }
+
+        @ParameterizedTest
+        @MethodSource("booleanValues")
+        void shouldIncludeHarFunktionsnedsattningValue(Boolean expectedValue) {
+            final var index = 1;
+
+            final var certificate = CertificateBuilder.create()
+                .addElement(QuestionHarFunktionsnedsattning.toCertificate(expectedValue, index, texts))
+                .build();
+
+            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
+
+            assertEquals(expectedValue, updatedCertificate.getHarFunktionsnedsattning());
+        }
+    }
+}
